fix: compensation and penalty calc

+++ HsDotNetProject/FreightMarket/Factories/TenderCompensationFactory.cs
@@ -6,11 +6,10 @@ public static class TenderCompensationFactory
 {
     private static readonly Random Random = new();
 
-    public static TenderCompensation From(TransportationGoods goods, int deliveryDurationInDays, DeliveryCharacteristics deliveryCharacteristics)
+    public static TenderCompensation From(TransportationGoods goods, double deliveryDurationInDays, DeliveryCharacteristics deliveryCharacteristics)
     {
         var multiplier = Random.Next(2);
-        // ReSharper disable once PossibleLossOfFraction
-        var bonusFactor  = 1 + (int)(0.2 + deliveryDurationInDays / deliveryCharacteristics.MaxDurationInDays) * multiplier;
+        var bonusFactor  = 1 + (0.2 + deliveryDurationInDays / deliveryCharacteristics.MaxDurationInDays) * multiplier;
         var compensation = goods.GoodPrice.MinPricePerTon * goods.WeightInTons * bonusFactor;
         return new TenderCompensation(compensation, goods.GoodPrice.Currency);
     }
